x_wing

Trait Decapsulator

pub trait Decapsulator {
    type Kem: Kem;

    // Required method
    fn encapsulation_key(&self) -> &<Self::Kem as Kem>::EncapsulationKey;
}

Decapsulator with an associated encapsulation key which can be used for encrypting shared keys that this decapsulator can decrypt.

Required Associated Types

type Kem: Kem

KEM algorithm this decapsulator is for.

Required Methods

fn encapsulation_key(&self) -> &<Self::Kem as Kem>::EncapsulationKey

Encapsulation key which can encrypt ciphertexts this decapsulator can decrypt.